# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "FacetBucket": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "share": {
            "title": "Share",
            "type": "number"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "count",
          "share"
        ],
        "title": "FacetBucket",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobDetail": {
        "properties": {
          "apply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apply Url"
          },
          "arrangement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrangement"
          },
          "canonical_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Title"
          },
          "canonical_title_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Title Slug"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "description_html": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Html"
          },
          "description_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Text"
          },
          "employment_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employment Type"
          },
          "experience_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Level"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/JobLocationOut"
            },
            "title": "Locations",
            "type": "array"
          },
          "posted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "salary": {
            "$ref": "#/components/schemas/SalaryOut"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/JobSectionOut"
            },
            "title": "Sections",
            "type": "array"
          },
          "source": {
            "$ref": "#/components/schemas/SourceOut"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "technologies": {
            "items": {
              "$ref": "#/components/schemas/JobTechnologyOut"
            },
            "title": "Technologies",
            "type": "array"
          },
          "title_raw": {
            "title": "Title Raw",
            "type": "string"
          },
          "years_of_experience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Of Experience"
          }
        },
        "required": [
          "id",
          "title_raw",
          "canonical_title",
          "canonical_title_slug",
          "company_name",
          "locations",
          "arrangement",
          "experience_level",
          "years_of_experience",
          "employment_type",
          "salary",
          "status",
          "posted_at",
          "last_seen_at",
          "apply_url",
          "description_text",
          "description_html",
          "sections",
          "technologies",
          "source"
        ],
        "title": "JobDetail",
        "type": "object"
      },
      "JobFilters": {
        "additionalProperties": false,
        "description": "Everything is optional; the empty filter means the national market.",
        "properties": {
          "arrangement": {
            "anyOf": [
              {
                "enum": [
                  "onsite",
                  "hybrid",
                  "remote"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrangement"
          },
          "experience_level": {
            "anyOf": [
              {
                "enum": [
                  "entry",
                  "mid",
                  "senior",
                  "staff_plus"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Level"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "q": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q"
          },
          "radius_miles": {
            "default": 50,
            "title": "Radius Miles",
            "type": "integer"
          },
          "salary_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Min"
          },
          "technologies": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Technologies",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "JobFilters",
        "type": "object"
      },
      "JobListItem": {
        "properties": {
          "apply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apply Url"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/JobLocationOut"
            },
            "title": "Locations",
            "type": "array"
          },
          "posted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "title_raw": {
            "title": "Title Raw",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title_raw",
          "company_name",
          "apply_url",
          "posted_at",
          "locations"
        ],
        "title": "JobListItem",
        "type": "object"
      },
      "JobListResponse": {
        "properties": {
          "facets": {
            "$ref": "#/components/schemas/SearchFacets"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "relaxations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RelaxationOption"
            },
            "title": "Relaxations",
            "type": "array"
          },
          "sort": {
            "enum": [
              "recency",
              "salary",
              "relevance"
            ],
            "title": "Sort",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "sort",
          "facets",
          "next_cursor"
        ],
        "title": "JobListResponse",
        "type": "object"
      },
      "JobLocationOut": {
        "properties": {
          "is_remote": {
            "title": "Is Remote",
            "type": "boolean"
          },
          "raw_text": {
            "title": "Raw Text",
            "type": "string"
          }
        },
        "required": [
          "raw_text",
          "is_remote"
        ],
        "title": "JobLocationOut",
        "type": "object"
      },
...
      "JobSectionOut": {
        "properties": {
          "end_offset": {
            "title": "End Offset",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "start_offset": {
            "title": "Start Offset",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "start_offset",
          "end_offset"
        ],
        "title": "JobSectionOut",
        "type": "object"
      },
      "JobTechnologyOut": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "evidence_snippet": {
            "title": "Evidence Snippet",
            "type": "string"
          },
          "evidence_start": {
            "title": "Evidence Start",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "requirement_level": {
            "title": "Requirement Level",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "category",
          "requirement_level",
          "confidence",
          "evidence_snippet",
          "evidence_start"
        ],
        "title": "JobTechnologyOut",
        "type": "object"
      },
...
      "RelaxationOption": {
        "description": "One honest way out of a zero-result search: what would change,\nand how many jobs that buys. Never applied server-side.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "filters": {
            "$ref": "#/components/schemas/JobFilters"
          },
          "kind": {
            "enum": [
              "radius",
              "salary",
              "technology"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "description",
          "filters",
          "count"
        ],
        "title": "RelaxationOption",
        "type": "object"
      },
...
      "SalaryOut": {
        "description": "Both representations, deliberately: raw for honest display\n('$60/hr'), normalized annual USD for anything comparative.",
        "properties": {
          "annual_max": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Max"
          },
          "annual_min": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Min"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "max_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Amount"
          },
          "min_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Amount"
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          }
        },
        "required": [
          "min_amount",
          "max_amount",
          "currency",
          "period",
          "annual_min",
          "annual_max"
        ],
        "title": "SalaryOut",
        "type": "object"
      },
...
      "SearchFacets": {
        "description": "Counts over the full matching set, not just the returned page.",
        "properties": {
          "arrangements": {
            "items": {
              "$ref": "#/components/schemas/FacetBucket"
            },
            "title": "Arrangements",
            "type": "array"
          },
          "experience_levels": {
            "items": {
              "$ref": "#/components/schemas/FacetBucket"
            },
            "title": "Experience Levels",
            "type": "array"
          }
        },
        "required": [
          "arrangements",
          "experience_levels"
        ],
        "title": "SearchFacets",
        "type": "object"
      },
...
      "SourceOut": {
        "properties": {
          "attribution_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution Text"
          },
          "attribution_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution Url"
          },
          "display_policy": {
            "title": "Display Policy",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "display_policy",
          "attribution_text",
          "attribution_url"
        ],
        "title": "SourceOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Xcelsior API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "description": "List matching active jobs with facet counts and keyset pagination.\n\nBoth the count and the rows come from the shared predicate \u2014 this is\nthe endpoint dashboard stats click through to, and its total must\nalways equal the number the user clicked.",
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "recency",
              "enum": [
                "recency",
                "salary",
                "relevance"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "description": "Opaque keyset cursor from a previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque keyset cursor from a previous page",
              "title": "Cursor"
            }
          },
          {
            "description": "Free-text query (websearch syntax)",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text query (websearch syntax)",
              "title": "Q"
            }
          },
          {
            "description": "Canonical title family slug",
            "in": "query",
            "name": "title",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Canonical title family slug",
              "title": "Title"
            }
          },
          {
            "description": "City slug, e.g. san-francisco-ca",
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "City slug, e.g. san-francisco-ca",
              "title": "Location"
            }
          },
          {
            "in": "query",
            "name": "radius_miles",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Radius Miles",
              "type": "integer"
            }
          },
          {
            "description": "Technology slugs (AND)",
            "in": "query",
            "name": "tech",
            "required": false,
            "schema": {
              "default": [],
              "description": "Technology slugs (AND)",
              "items": {
                "type": "string"
              },
              "title": "Tech",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "arrangement",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "onsite",
                    "hybrid",
                    "remote"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arrangement"
            }
          },
          {
            "in": "query",
            "name": "experience_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "entry",
                    "mid",
                    "senior",
                    "staff_plus"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Experience Level"
            }
          },
          {
            "description": "Annual USD floor",
            "in": "query",
            "name": "salary_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Annual USD floor",
              "title": "Salary Min"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}